ASoC: soc: pcm: add support for DPCM BE to BE connection
Some FEs can connect BE capture audio interfaces to BE playback audio
interfaces. This is accomplished by adding the dpcm_loopback flag in the
snd_soc_dai_link structure, allowing the machine driver to mention that the
FE (dynamic set) has loopback capability (capture->playback).
The FE will have its callbacks called twice, once for capture stream and
once for playback stream. The HW parameters for FE will not be set, since
the HW parameters are on the BEs.
The transfers are executed with minimal to none CPU intervention.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@microchip.com>