media: platform: dwc: dw-csi-plat: implement media controller
Implement media controller, pad links, and the bound for the subdevice.
DWC platform has a subdevice sensor that is connected on the first port.
Implement the possibility to access this subdevice.
The subdevice must be connected asynchronously.
Implement the notifier to have this subdev connected.
Implement start streaming video ops for the subdevice.
Adapt the code to linux 5.15.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>