clk: at91: sama7g5: Allow MCK1 to be exported and referenced in DT
This patch is taken on the dt branch because it fixes a compilation
error that is reproducible just on this branch and that was introduced
with:
d71954b84087 ("ARM: dts: at91: sama7g5: Add NAND support")
The error is not seen on linux-5.15-mchp because the
at91-5.15-trunk/base_clk branch was merged before at91-5.15-trunk/dt.
Add this patch here as well so that we don't break the build on the dt
branch.

MCK1 feeds the External Bus Interface (EBI). EBI's clock rate is used
to translate EBI's timmings to SMC timings, thus we need to handle MCK1
in the EBI driver. Allow MCK1 to be referenced as a PMC_TYPE_CORE clock
from phandle in DT.

Fixes: d71954b84087 ("ARM: dts: at91: sama7g5: Add NAND support")
Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>
Link: https://lore.kernel.org/r/20220111125310.902856-1-tudor.ambarus@microchip.com
Reviewed-by: Claudiu Beznea <claudiu.beznea@microchip.com>
Signed-off-by: Stephen Boyd <sboyd@kernel.org>
(cherry picked from commit a5ab04af49434aef532bf6cd4baa08a13665d608)
[ta: fixed compilation error on at91-5.15-trunk/dt]
Signed-off-by: Tudor Ambarus <tudor.ambarus@microchip.com>