ARM: dts: at91: sama7g5ek: enable pull-up on flexcom3 console lines
Flexcom3 is used as board console serial. There are no pull-ups on these
lines on the board. This means that if a cable is not connected (that has
pull-ups included), stray characters could appear on the console as the
floating pins voltage levels are interpreted as incoming characters.
To avoid this problem, enable the internal pull-ups on these lines.

Signed-off-by: Eugen Hristev <eugen.hristev@microchip.com>