riscv: dts: microchip: fix usage of fic clocks on mpfs
The fic clocks passed to the pcie controller and other peripherals in
the device tree are not the clocks they actually run on. The fics are
actually clock domain crossers & the clock config blocks output is the
mss/cpu side input. The peripherals are actually clocked by fixed
frequency clocks embedded in the fpga fabric. Update the device tree so
that these peripherals use the correct clocks.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>