rpmsg: miv: drop 16 bit shift from virtqueue id (vqid) and rename ihc include
All rpmsg/remoteproc drivers from NXP on the master and remote side
(rpmsg-lite) use a 16-bit shift on the virtqueue id. This is not
necessary on PolarFire SoC devices, therefore remove this shift and send
the index of the virtqueue directly as the payload as done by other
vendors.

Also, rename the include file from miv_ihc_message.h to miv_ihc.h to
make the header file more generic.

Note: The Mi-V RPMsg platform driver will be replaced with a Mi-V
remoteproc platform driver that will support rpmsg communication using
remoteproc virtio driver. In addition to this, it will support life
cycle management of a remote processor.

From a user point of view, all rpmsg Linux user space applications will
be compatible and remain unchanged.

The Mi-V Rpmsg stanza will be removed from the Icicle Kit device
tree in this release, since remoteproc will be use instead.

The Mi-V Rpmsg driver will remain in the Linux tree for one more release
to allow users to have this option as a fallback while they migrate over
to using remoteproc. However, it will be dropped from Linux in the next
release.

Users who still need to use the Mi-V RPMsg in this release will have to
add the rpmsg stanza back in the device tree and use the 2022.09 HSS and
2022.09 polarfire-soc-amp-examples.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Acked-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>