riscv: dts: microchip: drop label from memory regions
The memory nodes have both label properties and dt labels defined.
Drop the label property inside the memory nodes as it is not required
and is causing warnings with dtbs_check.

This fixes the following warning:

arch/riscv/boot/dts/microchip/mpfs-icicle-kit-context-a.dt.yaml: /: memory@80000000: 'label' does not match any of the regexes: 'pinctrl-[0-9]+'

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>