dt-bindings: riscv: document icicle reference design
The icicle kit reference design's v2022.10 release made some changes
to the memory map which adds the ability to use non-cached
aliases of cached memory addresses with DMA initiators & makes the
PCI controller to work again.

Make use of a new compatible to denote that this is not backwards
compatible.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Conor Dooley <conor.dooley@microchip.com>