fpga: microchip-auto-update: return count, not zero from debugfs writes
As this is just being used to trigger the bitstream loading process, the
number of chars written has no meaning here. Returning 0 is not correct,
as it leads the filesystem to think that the write has not completed.
As a result, the write loops ad infinitum.
Return count instead, so that the filesystem knows the write has
completed.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>