mfd: atmel-hlcdc: set prepare and enable LVDS PLL clock for LVDS display
The XLCDC IP in sam9x7 supports DSI, parallel RGB and LVDS Display.
sys_clk (Generic clock) of atmel-hlcdc is used with DSI and Parallel
RGB and LVDS PLL is used with LVDS display.

Set the LVDS PLL clock rate by querying the panel's pixel clock frequency.

prepare and enable the lvds_pll_clk, if it is configured in DT.

Signed-off-by: Manikandan Muralidharan <manikandan.m@microchip.com>
Signed-off-by: Dharma Balasubiramani <dharma.b@microchip.com>