ARM: dts: at91: sam9x60: fix the SOC detection
Defining the dbgu compatible strings in the UART submodule of the
flexcomm 5 gives the below error log,
AT91: Could not find matching SoC description
This error arises due to defining the dbgu compatible strings to
the flexcom usart which is not valid because there is only one debug unit
in the sam9x60 SOC and it has the chipid register. The dbgu compatible
strings are valid only for debug usart and not valid for flexcom usart.
Thus removing the dbgu compatible strings in the UART submodule of the
flexcom for the proper SOC detection.

Fixes: c016adebe2fe (ARM: dts: at91: sam9x60: Move flx5 definitions
in the SoC dtsi)
Signed-off-by: Durai Manickam KR <durai.manickamkr@microchip.com>