riscv: dts: microchip: add missing properties for fabric-bus 4
ranges is a required property for simple-bus, so add an empty one as
there is no translation here.
The bus has a unit address, so it requires a reg property too. The only
thing actually on this bus is the athena crypto, so the reg property
here matches that of the crypto 1:1.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>